Extract boltdb JobDB implementation to new package
