+++ job/storage/boltdb/boltdb.go
@@ -1,4 +1,4 @@
-package storage
+package boltdb
 
 import (
 	"bytes"
