+++ job/storage/boltdb/boltdb_test.go
@@ -1,4 +1,4 @@
-package storage
+package boltdb
 
 import (
 	"testing"
