+++ main.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/ajvb/kala/api"
 	"github.com/ajvb/kala/job"
-	"github.com/ajvb/kala/job/storage"
+	"github.com/ajvb/kala/job/storage/boltdb"
 	"github.com/ajvb/kala/utils/logging"
 
 	"github.com/codegangsta/cli"
@@ -63,7 +63,7 @@ func main() {
 					connectionString = parsedPort
 				}
 
-				db := storage.GetBoltDB(c.String("boltpath"))
+				db := boltdb.GetBoltDB(c.String("boltpath"))
 
 				// Create cache
 				cache := job.NewMemoryJobCache(db, DefaultPersistEvery)
